Aplicando correcciones
