+++ frontend/server/src/Controllers/Session.php
@@ -2,6 +2,24 @@
 
 namespace OmegaUp\Controllers;
 
+class ScopedFacebook {
+    /** @var \OmegaUp\ScopedSession */
+    public $scopedSession;
+    /** @var \Facebook\Facebook */
+    public $facebook;
+
+    public function __construct() {
+        require_once 'libs/third_party/facebook-php-graph-sdk/src/Facebook/autoload.php';
+
+        $this->scopedSession = new \OmegaUp\ScopedSession();
+        $this->facebook = new \Facebook\Facebook([
+            'app_id' => OMEGAUP_FB_APPID,
+            'app_secret' => OMEGAUP_FB_SECRET,
+            'default_graph_version' => 'v2.5',
+        ]);
+    }
+}
+
 /**
  * Session controller handles sessions.
  * @psalm-type AssociatedIdentity=array{default: bool, username: string}
@@ -24,7 +42,7 @@ class Session extends \OmegaUp\Controllers\Controller {
     }
 
     public static function getFacebookLoginUrl(): string {
-        $scopedFacebook = new \OmegaUp\ScopedFacebook();
+        $scopedFacebook = new ScopedFacebook();
         $helper = $scopedFacebook->facebook->getRedirectLoginHelper();
         return $helper->getLoginUrl(OMEGAUP_URL . '/login?fb', ['email']);
     }
@@ -385,13 +403,149 @@ class Session extends \OmegaUp\Controllers\Controller {
 
     /**
      * Logs in via Facebook API.
+     */
+    public static function loginViaFacebook(): void {
+        // Mostly taken from
+        // https://developers.facebook.com/docs/php/howto/example_facebook_login
+        $scopedFacebook = new ScopedFacebook();
+        $helper = $scopedFacebook->facebook->getRedirectLoginHelper();
+        try {
+            $accessToken = $helper->getAccessToken();
+        } catch (\Facebook\Exceptions\FacebookResponseException $e) {
+            $errorMessage = $e->getMessage();
+            self::$log->error("Graph returned an error: {$errorMessage}");
+            throw $e;
+        } catch (\Facebook\Exceptions\FacebookSDKException $e) {
+            $errorMessage = $e->getMessage();
+            self::$log->error(
+                "Facebook SDK returned an error: {$errorMessage}"
+            );
+            throw $e;
+        }
+
+        if (is_null($accessToken) && !is_null($helper->getError())) {
+            $errorDescription = $helper->getErrorDescription();
+            self::$log->error(
+                "Unable to login via Facebook: {$errorDescription}"
+            );
+            throw new \OmegaUp\Exceptions\InvalidParameterException(
+                'parameterNotFound',
+                'token'
+            );
+        }
+
+        try {
+            $fbResponse = $scopedFacebook->facebook->get(
+                '/me?fields=name,email',
+                $accessToken
+            );
+        } catch (\Facebook\Exceptions\FacebookResponseException $e) {
+            self::$log->error("Unable to login via Facebook: {$e}");
+            throw $e;
+        } catch (\Facebook\Exceptions\FacebookSDKException $e) {
+            self::$log->error("Unable to login via Facebook: {$e}");
+            throw $e;
+        }
+
+        $fbUserProfile = $fbResponse->getGraphUser();
+        self::$log->info('User is logged in via facebook !!');
+        if (is_null($fbUserProfile->getEmail())) {
+            self::$log->error('Facebook email empty');
+            throw new \OmegaUp\Exceptions\InvalidParameterException(
+                'loginFacebookEmptyEmailError',
+                'error'
+            );
+        }
+
+        \OmegaUp\Controllers\Session::thirdPartyLogin(
+            'Facebook',
+            strval($fbUserProfile->getEmail()),
+            $fbUserProfile->getName()
+        );
+
+        if (\OmegaUp\Controllers\Session::currentSessionAvailable()) {
+            $redirectUrl = self::getRedirectUrl();
+            header("Location: {$redirectUrl}");
+            die();
+        }
+    }
+
+    public static function loginViaLinkedIn(
+        string $code,
+        string $state,
+        string $redirect
+    ): void {
+        try {
+            $li = self::getLinkedInInstance($redirect);
+            $authToken = $li->getAuthToken($code, $state);
+            $profile = $li->getProfileInfo($authToken);
+            $redirect = $li->extractRedirect($state);
+
+            \OmegaUp\Controllers\Session::thirdPartyLogin(
+                'LinkedIn',
+                $profile['emailAddress'],
+                "{$profile['firstName']} {$profile['lastName']}"
+            );
+            if (\OmegaUp\Controllers\Session::currentSessionAvailable()) {
+                $redirectUrl = self::getRedirectUrl($redirect);
+                header("Location: {$redirectUrl}");
+                die();
+            }
+        } catch (\OmegaUp\Exceptions\ApiException $e) {
+            self::$log->error("Unable to login via LinkedIn: $e");
+            throw $e;
+        }
+    }
+
+    private static function getLinkedInInstance(
+        ?string $redirect = null
+    ): \OmegaUp\LinkedIn {
+        return new \OmegaUp\LinkedIn(
+            OMEGAUP_LINKEDIN_CLIENTID,
+            OMEGAUP_LINKEDIN_SECRET,
+            OMEGAUP_URL . '/login?linkedin',
+            $redirect
+        );
+    }
+
+    public static function getLinkedInLoginUrl(): string {
+        return self::getLinkedInInstance()->getLoginUrl();
+    }
+
+    private static function getRedirectUrl(?string $url = null): string {
+        $defaultRedirectUrl = '/profile/';
+        if (is_null($url)) {
+            return $defaultRedirectUrl;
+        }
+        $redirectParsedUrl = parse_url($url);
+        // If a malformed URL is given, don't redirect.
+        if ($redirectParsedUrl === false) {
+            return $defaultRedirectUrl;
+        }
+        // Just the path portion of the URL was given.
+        if (
+            empty($redirectParsedUrl['scheme']) ||
+            empty($redirectParsedUrl['host'])
+        ) {
+            $path = $redirectParsedUrl['path'] ?? '';
+            return $path !== '/logout/' ? $url : $defaultRedirectUrl;
+        }
+        $redirectUrl = "{$redirectParsedUrl['scheme']}://{$redirectParsedUrl['host']}";
+        if (isset($redirectParsedUrl['port'])) {
+            $redirectUrl .= ":{$redirectParsedUrl['port']}";
+        }
+        return $redirectUrl === OMEGAUP_URL ? $url : $defaultRedirectUrl;
+    }
+
+    /**
+     * Logs in via Facebook API.
      *
      * @return array<string, mixed>
      */
-    public static function LoginViaFacebook(): array {
+    public static function LoginViaFacebookDeprecated(): array {
         // Mostly taken from
         // https://developers.facebook.com/docs/php/howto/example_facebook_login
-        $scopedFacebook = new \OmegaUp\ScopedFacebook();
+        $scopedFacebook = new ScopedFacebook();
         $helper = $scopedFacebook->facebook->getRedirectLoginHelper();
         try {
             $accessToken = $helper->getAccessToken();
@@ -564,7 +718,7 @@ class Session extends \OmegaUp\Controllers\Controller {
         }
     }
 
-    public static function getLinkedInInstance(): \OmegaUp\LinkedIn {
+    public static function getLinkedInInstanceDeprecated(): \OmegaUp\LinkedIn {
         return new \OmegaUp\LinkedIn(
             OMEGAUP_LINKEDIN_CLIENTID,
             OMEGAUP_LINKEDIN_SECRET,
@@ -572,14 +726,15 @@ class Session extends \OmegaUp\Controllers\Controller {
             \OmegaUp\Request::getRequestVar('redirect')
         );
     }
-    public static function getLinkedInLoginUrl(): string {
-        return self::getLinkedInInstance()->getLoginUrl();
+
+    public static function getLinkedInLoginUrlDeprecated(): string {
+        return self::getLinkedInInstanceDeprecated()->getLoginUrl();
     }
 
     /**
      * @return array<string, mixed>
      */
-    public static function LoginViaLinkedIn(): array {
+    public static function LoginViaLinkedInDeprecated(): array {
         $code = \OmegaUp\Request::getRequestVar('code');
         $state = \OmegaUp\Request::getRequestVar('state');
         if (empty($code) || empty($state)) {
